fix: avoid leaking `@types/node` into client code
