+++ packages/wxt/src/core/resolve-config.ts
@@ -361,7 +361,6 @@ async function getUnimportOptions(
   ];
 
   const defaultOptions: WxtResolvedUnimportOptions = {
-    imports: [{ name: 'fakeBrowser', from: 'wxt/testing' }],
     presets: [
       {
         from: 'wxt/browser',
