+++ packages/wxt/src/utils/define-background.ts
@@ -1,5 +1,5 @@
 /** @module wxt/utils/define-background */
-import type { BackgroundDefinition } from '../types';
+import type { BackgroundDefinition } from '../client-types';
 
 export function defineBackground(main: () => void): BackgroundDefinition;
 export function defineBackground(
