+++ packages/wxt/src/utils/define-content-script.ts
@@ -1,5 +1,5 @@
 /** @module wxt/utils/define-content-script */
-import type { ContentScriptDefinition } from '../types';
+import type { ContentScriptDefinition } from '../client-types';
 
 export function defineContentScript(
   definition: ContentScriptDefinition,
