+++ packages/wxt/src/utils/define-wxt-plugin.ts
@@ -1,5 +1,5 @@
 /** @module wxt/utils/define-wxt-plugin */
-import type { WxtPlugin } from '../types';
+import type { WxtPlugin } from '../client-types';
 
 export function defineWxtPlugin(plugin: WxtPlugin): WxtPlugin {
   return plugin;
